requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
django-crispy-forms
djangorestframework
dj-rest-auth
django-allauth
django-rest-hooks
